clean(core): fixes after review
