+++ core/src/main/java/io/kestra/core/repositories/TenantMigrationInterface.java
@@ -2,6 +2,6 @@ package io.kestra.core.repositories;
 
 public interface TenantMigrationInterface {
 
-    void migrateTenant(boolean dryRun);
+    void migrateTenant(String tenantId, boolean dryRun);
 
 }
